i-nergy-clustering/requirements.txt
grpcio==1.59.2
grpcio-tools==1.59.2
mlflow==2.8.1
pandas==2.1.3
protobuf==4.25.1
PyYAML==6.0.1
